added minor thing to make it start

+++ src/main/java/com/pupiq/restfordiploma/repository/UserRoleRepository.java
@@ -3,6 +3,6 @@ package com.pupiq.restfordiploma.repository;
 import com.pupiq.restfordiploma.model.UserRole;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRoleRepository extends JpaRepository<UserRole, Long> {
+public interface UserRoleRepository extends JpaRepository<UserRole, Integer> {
 
 }
